# AIM-OpenAPI.yaml
openapi: 3.0.3
...
info:
  title: Armory Inventory Management APIs
  description: APIs for managing AIM (Armory Inventory Management)
  version: 1.0.0
...
paths:
...
  /auth/login:
    post:
      summary: Super Admin Login
      operationId: superAdminLogin
      tags:
        - SuperAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuperAdminRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuperAdminResponse'
              example:
                token: "jwttoken"
                tokenExpiry: "2023-04-01T12:00:00Z"
                role: "SuperAdmin"
                id: 12345
                email: "john.doe@example.com"
                name: "John Doe"
                phoneNumber: "+1234567890"
                createdAt: "2023-04-01T12:00:00Z"
                updatedAt: "2023-04-01T12:00:00Z"
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuperAdminResponse'
              example:
                message: "Email and password are required"
                errorCode: "BAD_REQUEST"

        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuperAdminResponse'
              example:
                message: "Invalid email or password"
                errorCode: "UNAUTHORIZED"

        '403':
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuperAdminResponse'
              example:
                message: "Account is disabled"
                errorCode: "FORBIDDEN"

        '404':
          description: Super Admin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuperAdminResponse'
              example:
                message: "Super Admin not found"
                errorCode: "NOT_FOUND"


  #=======================================Organization APIs===========================================================
...
  /category:
    get:
      summary: Get all categories
      operationId: getAllCategories
      tags:
        - Category
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: A list of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDto'
        '500':
          description: Internal server error
    post:
      summary: Create a new category
      operationId: createCategory
      tags:
        - Category
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDtoInsert'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDtoInsert'
        '400':
          description: Bad request - Missing or invalid data
        '409':
          description: Category with the same name already exists
        '500':
          description: Internal server error
  /category/{categoryId}:
    get:
      summary: Get category by ID
      operationId: getCategoryById
      tags:
        - Category
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID of the category to fetch
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Category details by ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
        '404':
          description: Category not found
        '500':
          description: Internal server error

    put:
      summary: Update category by ID
      operationId: updateCategory
      tags:
        - Category
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID of the category to update
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDtoInsert'
      responses:
        '200':
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDtoInsert'
        '404':
          description: Category not found
        '400':
          description: Bad request - Missing or invalid data
        '500':
          description: Internal server error

    delete:
      summary: Delete category by ID
      operationId: deleteCategory
      tags:
        - Category
      parameters:
        - name: categoryId
          in: path
          required: true
          description: ID of the category to delete
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Category deleted successfully
        '404':
          description: Category not found
        '500':
          description: Internal server error

  #=======================================Inventory APIs===========================================================

  # Get all inventories
...
  /inventory/category/{categoryId}:
    get:
      summary: Get inventories by categoryId
      operationId: getInventoriesByCategoryId
      tags:
        - Inventory
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Category ID to fetch associated inventories
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'

      responses:
        '200':
          description: List of inventories for the specified category ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryDto'
        '404':
          description: No inventory found for the category
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "No inventory found for the category"
        '500':
          description: Internal server error

#  # Get inventory by asset item number
#  /inventory/assetItem/{assetItemNumber}:
#    get:
#      summary: Get inventory by asset item number
#      operationId: getInventoryByAssetItemNumber
#      tags:
#        - Inventory
#      parameters:
#        - name: assetItemNumber
#          in: path
#          required: true
#          description: Asset item number of the inventory to fetch
#          schema:
#            type: string
#        - $ref: '#/components/parameters/X-Tenant-ID'
#
#      responses:
#        '200':
#          description: Inventory details by asset item number
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/InventoryDto'
#        '404':
#          description: Inventory not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
#=============================================================================================================




#=====================================================COMPONENTS===========================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    X-Tenant-ID:
      name: X-Tenant-ID
      in: header
      description: Tenant identifier for multi-tenancy.
      required: true
      schema:
        type: string
#    Authorization:
#      name: Authorization
#      in: header
#      description: "Bearer token for authentication. Format: `Bearer <token>`"
#      required: true
#      schema:
#        type: string
  schemas:
...
    SuperAdminRequest:
      type: object
      properties:
        email:
          type: string
          example: "email@example.com"
        password:
          type: string
          example: "password"
      required:
        - email
        - password
    SuperAdminResponse:
      oneOf:
        - $ref: '#/components/schemas/SuperAdminSuccessResponse'
        - $ref: '#/components/schemas/SuperAdminErrorResponse'
    SuperAdminSuccessResponse:
      type: object
      properties:
        token:
          type: string
          example: "***.."
        tokenExpiry:
          type: string
          format: date-time
          description: The timestamp when the SuperAdmin was created.
        role:
          type: string
          example: "role"
        id:
          type: integer
          format: int64
          description: The unique identifier of the SuperAdmin.
        email:
          type: string
          example: "email@example.com"
        name:
          type: string
          example: "John Doe"
        phoneNumber:
          type: string
          example: "+1234567890"
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the SuperAdmin was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the SuperAdmin was last updated.
    SuperAdminErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Error message here"
      required:
        - error
#===============================================Organization DTOs===================================================================
...
    ManufacturerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the manufacturer
        name:
          type: string
          description: Name of the manufacturer
        status:
          type: string
          description: Status of the manufacturer (e.g., active, inactive)
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the manufacturer was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the manufacturer was last updated
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
          description: List of categories associated with the manufacturer.
      required:
        - name

    #Category Schema
    CategoryDto:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          description: The unique identifier for the category.
        manufacturers: # Many-to-many relationship with ManufacturerDto
          type: array
          items:
            $ref: '#/components/schemas/ManufacturerDto'
          description: List of manufacturers associated with the category.
        categoryName:
          type: string
          description: The name of the category.
        categoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing various details about the category (e.g., model, type, etc.).
          example:
            model: "AR-15"
            type: "Rifle"
            required_training: "Basic"
            required_qualification: "Marksman"
            caliber: "5.56mm"
            length: "36 inches"
            inspection_months: "12"
            maintenance_months: "6"
            description: "A standard assault rifle."

        inventories: # One-to-many association with InventoryDto
          type: array
          items:
            $ref: '#/components/schemas/InventoryDto'
          description: List of inventories associated with the category.
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the category was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the category was last updated.
      required:
        - categoryName
        - categoryDetails
    CategoryDtoInsert:
      type: object
      properties:
        categoryName:
          type: string
          description: The name of the category.
        categoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing various details about the category (e.g., model, type, etc.).
          example:
            model: "AR-15"
            type: "Rifle"
            required_training: "Basic"
            required_qualification: "Marksman"
            caliber: "5.56mm"
            length: "36 inches"
            inspection_months: "12"
            maintenance_months: "6"
            description: "A standard assault rifle."
        manufacturerIds:
          type: array
          items:
            type: integer
            format: int64
          example: [ 1, 2, 3 ]
      required:
        - categoryName
        - categoryDetails
        - manufacturerIds



    #Inventory Dto
    InventoryDto:
      type: object
      properties:
        inventoryId:
          type: integer
          format: int64
          description: The unique identifier for the inventory item.
#        categoryId:
#          type: integer
#          format: int64
#          description: The unique identifier of the associated category.
        category: # Many-to-one association with CategoryDto
          $ref: '#/components/schemas/CategoryDto'
          description: The associated category for the inventory item.
#        assetItemNumber:
#          type: string
#          description: The unique asset item number for the inventory.
        serialNo:
          type: string
          description: The serial number of the inventory item.
        inventoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing varisous details about the inventory item (e.g., lot number, rebuild date, etc.).
          example:
            model: "AQW"
            assetItemNumber: "AS123"
            site: "site1"
            nfa_category: "Rifle"
            description: "Brand new rifle, fully assembled and tested."
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was last updated.
      required:
        - serialNo
        - categoryId


    #Inventory Dto
...